// frontend/src/types/ban-doc.ts
export type BanDoc = {
  maBanDoc: string
  hoTen: string
  soThe: string
  email: string
  dienThoai: string
  hanThe: string
  trangThaiThe: string
  duNo: number
}

export type CreateBanDocPayload = {
  hoTen: string
  soThe: string
  email: string
  dienThoai: string
  hanThe: string
  trangThaiThe: string
  duNo: number
}

export type UpdateBanDocPayload = {
  hoTen: string
  soThe: string
  email: string
  dienThoai: string
  hanThe: string
  trangThaiThe: string
  duNo: number
}